10/5/19 - Created sprites and the queue
10/5/19
Changes implemented
-You can see the next block and the rest of the queue

Basically, I got rid of the old code that drew rectangles and put in some new code which loads picture files
as textures and draws the textures to the screen

The queue takes up the entire length of the score tab which makes me think that perhaps I should modify the code
such that the queue does not take up the entire length of the score tab. Especially since I will want to show the 
hold queue as well. I could also lower the amount of blocks that are shown. Maybe to 4? So you can only know the
next block and 4 subsequent blocks?

The sprites are each 100 x 100
They are the same color as their corresponding blocks in-game.
I is 80 across, 20 high
O is 40 x 40
All the other blocks used a 60 x 40 template

Changes that should be made...
-Adding functionality to the options menu...
-Ghost 
-Holding blocks/Hold block
-Adding freezing animations
-Score table
-Scoring/Stat-tracking...